refactor soem code
